ChebPade

Documentation for ChebPade.

ChebPade.chebpade

ChebPade.clenshawlord

ChebPade.extendrandn

ChebPade.maehly

ChebPade.chebpade — Method

chebpade(f::Fun{<:Chebyshev},m::Integer,n::Integer;method=:clenshawlord)

computes the (m,n) Chebyshev-Pade approximation to f using method :clenshawlord or :maehly

ChebPade.clenshawlord — Method

clenshawlord(F::Fun{<:Chebyshev}, m::Integer, n::Integer)

computes the (m,n) Chebyshev-Pade approximant to F using the Clenshaw-Lord method

ChebPade.extendrandn — Method

extendrandn(x::AbstractVector{T},n::Integer)

returns a vector of length n, where the trailing entries are given by randn scaled by eps(float(real(T)))

ChebPade.maehly — Method

maehly(F::Fun{<:Chebyshev}, m::Integer, n::Integer)

computes the (m,n) Chebyshev-Pade approximant to F using the Maehly method